rename pkg logger/log to logger/printer
