+++ pkg/cluster/executor/local_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	"github.com/pingcap/tiup/pkg/cluster/ctxt"
-	logprinter "github.com/pingcap/tiup/pkg/logger/log"
+	logprinter "github.com/pingcap/tiup/pkg/logger/printer"
 	"github.com/stretchr/testify/require"
 )
